dom/media/webaudio/test/test_scriptProcessorNodeNotConnected.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioBufferSourceNode: should not fire audioprocess if not connected.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("This test needs to wait a while to ensure that a given event does not happen.");
addLoadEvent(function() {
  var context = new AudioContext();

  var sp = context.createScriptProcessor(2048, 2, 2);
  sp.onaudioprocess = function(e) {
    ok(false, "Should not call onaudioprocess if the node is not connected.");
    sp.onaudioprocess = null;
    SimpleTest.finish();
  };
  setTimeout(function() {
    console.log(sp.onaudioprocess);
    if (sp.onaudioprocess) {
      ok(true, "onaudioprocess not fired.");
      SimpleTest.finish();
    }
  }, 4000);
});
</script>
</pre>
</body>
</html>
